% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hospital_stays.R
\name{build_da_conditions}
\alias{build_da_conditions}
\title{Construit les conditions pour extraire les diagnostics associés.}
\usage{
build_da_conditions(cim10_codes_starts_with = NULL)
}
\arguments{
\item{cim10_codes_starts_with}{character vector Les codes CIM10 cibles des
diagnostics à extraire.}
}
\value{
character Les conditions pour extraire les diagnostics associés.
}
\description{
Cette fonction permet de construire les conditions pour extraire les
diagnostics associés.
}
\examples{
build_da_conditions(c("A00", "B00"))
}
