% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hospital_stays.R
\name{build_dp_dr_conditions}
\alias{build_dp_dr_conditions}
\title{Construit les conditions pour extraire les diagnostics principaux et reliés.}
\usage{
build_dp_dr_conditions(cim10_codes = NULL, include_dr = FALSE)
}
\arguments{
\item{cim10_codes}{character vector Les codes CIM10 cibles des
diagnostics à extraire.}

\item{include_dr}{logical Indique si les diagnostics reliés doivent être
ajoutés dans les conditions. Si TRUE, la recherche dans les diagnostics
principaux des codes CIM10 cibles est ajoutée dans les conditions. Si FALSE,
les codes CIM10 cibles sont recherchés seulement pour les diagnostics
principaux.}
}
\value{
character Les conditions pour extraire les diagnostics principaux et
reliés.
}
\description{
Cette fonction permet de construire les conditions pour extraire les
diagnostics principaux et reliés.
}
\examples{
\dontrun{
build_dp_dr_conditions(c("A00", "B00"), include_dr = TRUE)
build_dp_dr_conditions(c("A00", "B00"), include_dr = FALSE)
}
}
\keyword{internal}
