% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{COLNAMES_DCIR_KEYS}
\alias{COLNAMES_DCIR_KEYS}
\title{Neufs clés de jointure du DCIR}
\format{
An object of class \code{character} of length 9.
}
\usage{
COLNAMES_DCIR_KEYS
}
\value{
Character vector. Neufs clés de jointure du DCIR : table principale ER_PRS_F et \href{https://documentation-snds.health-data-hub.fr/snds/fiches/tables_affinees.html#les-tables-affinees#les-tables-affinees}{tables affinées}.
}
\description{
Neufs clés de jointure du DCIR
}
\seealso{
Other constants: 
\code{\link{IS_PORTAIL}}
}
\concept{constants}
\keyword{datasets}
