% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect_duckdb}
\alias{connect_duckdb}
\title{Initialisation de la connexion à la base de données duckdb.}
\usage{
connect_duckdb()
}
\value{
dbConnection Connexion à la base de données duckdb
}
\description{
Utilisation pour le testing uniquement. Si le code s'exécute en dehors du
portail, il faut initier une connexion duckdb pour effectuer les tests.
}
\seealso{
Other utils: 
\code{\link{connect_oracle}()},
\code{\link{constants_snds_tools}()},
\code{\link{create_table_from_query}()},
\code{\link{gather_table_stats}()},
\code{\link{get_first_non_archived_year}()},
\code{\link{insert_into_table_from_query}()},
\code{\link{retrieve_all_psa_from_idt}()},
\code{\link{retrieve_all_psa_from_psa}()}
}
\concept{utils}
