% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect_oracle}
\alias{connect_oracle}
\title{Initialisation de la connexion à la base de données.}
\usage{
connect_oracle()
}
\value{
dbConnection Connexion à la base de données oracle
}
\description{
Initialisation de la connexion à la base de données.
}
\seealso{
Other utils: 
\code{\link{connect_duckdb}()},
\code{\link{constants_snds}()},
\code{\link{create_table_from_query}()},
\code{\link{gather_table_stats}()},
\code{\link{get_first_non_archived_year}()},
\code{\link{insert_into_table_from_query}()},
\code{\link{retrieve_all_psa_from_idt}()},
\code{\link{retrieve_all_psa_from_psa}()}
}
\concept{utils}
