% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants_snds}
\alias{constants_snds}
\title{Stocke des constantes pour sndsTools}
\usage{
constants_snds()
}
\value{
A list of constants for sndsTools.
}
\description{
Stocke des constantes pour sndsTools
}
\seealso{
Other utils: 
\code{\link{connect_duckdb}()},
\code{\link{connect_oracle}()},
\code{\link{create_table_from_query}()},
\code{\link{gather_table_stats}()},
\code{\link{get_first_non_archived_year}()},
\code{\link{insert_into_table_from_query}()},
\code{\link{parallelize_query_by_flx_month}()},
\code{\link{retrieve_all_psa_from_idt}()},
\code{\link{retrieve_all_psa_from_psa}()}
}
\concept{utils}
