% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants_snds_tools}
\alias{constants_snds_tools}
\title{Stocke des constantes pour SndsTools}
\usage{
constants_snds_tools()
}
\value{
A list of constants for SNDs tools.
}
\description{
Stocke des constantes pour SndsTools
}
\seealso{
Other utils: 
\code{\link{connect_duckdb}()},
\code{\link{connect_oracle}()},
\code{\link{create_table_from_query}()},
\code{\link{gather_table_stats}()},
\code{\link{get_first_non_archived_year}()},
\code{\link{insert_into_table_from_query}()},
\code{\link{retrieve_all_psa_from_idt}()},
\code{\link{retrieve_all_psa_from_psa}()}
}
\concept{utils}
