% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{create_mock_database}
\alias{create_mock_database}
\title{Configurer une base de données DuckDB avec toutes les tables factices}
\usage{
create_mock_database(
  n_patients = 100,
  year = 2020,
  start_date = as.Date("2020-01-01"),
  end_date = as.Date("2020-12-31")
)
}
\arguments{
\item{n_patients}{Nombre de patients à générer}

\item{year}{Année pour les tables MCO}

\item{start_date}{Date de début}

\item{end_date}{Date de fin}
}
\value{
Connexion DuckDB avec toutes les tables chargées
}
\description{
Configurer une base de données DuckDB avec toutes les tables factices
}
