% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{create_mock_er_ete_f}
\alias{create_mock_er_ete_f}
\title{Créer des données factices pour ER_ETE_F (Actes externes)}
\usage{
create_mock_er_ete_f(er_prs_f)
}
\arguments{
\item{er_prs_f}{data.frame des prescriptions}
}
\value{
data.frame avec données d'actes externes
}
\description{
Créer des données factices pour ER_ETE_F (Actes externes)
}
