% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{create_mock_er_pha_f}
\alias{create_mock_er_pha_f}
\title{Créer des données factices pour les médicaments délivrés (ER_PHA_F)}
\usage{
create_mock_er_pha_f(er_prs_f)
}
\arguments{
\item{er_prs_f}{data.frame des prescriptions}
}
\value{
data.frame avec les données des médicaments
}
\description{
Créer des données factices pour les médicaments délivrés (ER_PHA_F)
}
