% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{create_mock_er_prs_f}
\alias{create_mock_er_prs_f}
\title{Créer des données factices pour les délivrances de médicaments (ER_PRS_F)}
\usage{
create_mock_er_prs_f(
  patients_ids = create_mock_patients_ids(),
  start_date = as.Date("2020-01-01"),
  end_date = as.Date("2020-12-31"),
  n_dispenses = 5
)
}
\arguments{
\item{patients_ids}{data.frame avec BEN_IDT_ANO et BEN_NIR_PSA}

\item{start_date}{Date de début}

\item{end_date}{Date de fin}

\item{n_dispenses}{Nombre de délivrances par patient (moyenne)}
}
\value{
data.frame avec les données de délivrances
}
\description{
Créer des données factices pour les délivrances de médicaments (ER_PRS_F)
}
