% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{create_mock_ir_ben_r}
\alias{create_mock_ir_ben_r}
\title{Créer des données factices pour IR_BEN_R (référentiel bénéficiaires)}
\usage{
create_mock_ir_ben_r(patients_ids = create_mock_patients_ids())
}
\arguments{
\item{patients_ids}{data.frame avec BEN_IDT_ANO et BEN_NIR_PSA}
}
\value{
data.frame avec toutes les colonnes du référentiel bénéficiaires
}
\description{
Créer des données factices pour IR_BEN_R (référentiel bénéficiaires)
}
