% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{create_mock_ir_imb_r}
\alias{create_mock_ir_imb_r}
\title{Créer des données factices pour les ALD (IR_IMB_R)}
\usage{
create_mock_ir_imb_r(
  patients_ids = create_mock_patients_ids(),
  start_date = as.Date("2020-01-01"),
  end_date = as.Date("2020-12-31")
)
}
\arguments{
\item{patients_ids}{data.frame avec BEN_IDT_ANO et BEN_NIR_PSA}

\item{start_date}{Date de début}

\item{end_date}{Date de fin}
}
\value{
data.frame avec les données d'ALD
}
\description{
Créer des données factices pour les ALD (IR_IMB_R)
}
