% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{create_mock_ir_pha_r}
\alias{create_mock_ir_pha_r}
\title{Créer des données factices pour le référentiel médicaments (IR_PHA_R)}
\usage{
create_mock_ir_pha_r()
}
\value{
data.frame avec les codes CIP13 et ATC
}
\description{
Créer des données factices pour le référentiel médicaments (IR_PHA_R)
}
