% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{create_mock_mco_tables}
\alias{create_mock_mco_tables}
\title{Créer des données factices pour les séjours hospitaliers MCO (tables B, C, D, UM)}
\usage{
create_mock_mco_tables(
  patients_ids = create_mock_patients_ids(),
  year = 2020,
  start_date = as.Date("2020-01-01"),
  end_date = as.Date("2020-12-31")
)
}
\arguments{
\item{patients_ids}{data.frame avec BEN_IDT_ANO et BEN_NIR_PSA}

\item{year}{Année des séjours}

\item{start_date}{Date de début}

\item{end_date}{Date de fin}
}
\value{
list avec les 4 tables MCO (B, C, D, UM)
}
\description{
Créer des données factices pour les séjours hospitaliers MCO (tables B, C, D, UM)
}
