% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{create_mock_patients_ids}
\alias{create_mock_patients_ids}
\title{Fonctions pour créer des données factices pour le tutoriel sndsTools}
\usage{
create_mock_patients_ids(n_patients = 100)
}
\arguments{
\item{n_patients}{Nombre de patients à générer}
}
\value{
data.frame avec BEN_IDT_ANO et BEN_NIR_PSA
}
\description{
Ce fichier contient des fonctions pour générer des données d'exemple
basées sur les structures utilisées dans les tests du package sndsTools.
}
\author{
Matthieu Doutreligne
Créer des données factices pour les patients
}
