% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_synthetic_snds}
\alias{download_synthetic_snds}
\title{Télécharge et zip les données synthétiques du SNDS}
\usage{
download_synthetic_snds()
}
\value{
Le chemin vers le répertoire contenant le fichier zip contenant les
données.
}
\description{
Télécharge \href{https://www.data.gouv.fr/fr/datasets/donnees-synthetiques-de-la-base-principales-du-systeme-national-des-donnees-de-sante/}{les données fictives du SNDS}
disponible sur datagouv.fr. Puis, zip dans une seul fichier
\code{synthetic_data.zip} toutes les tables des données fictives. Sauvegarde ce
fichier zip dans le répertoire \code{inst/extdata} du package.
}
