% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_consultations_erprsf.R
\name{extract_consultations_erprsf}
\alias{extract_consultations_erprsf}
\title{Extraction des consultations dans le DCIR.}
\usage{
extract_consultations_erprsf(
  start_date,
  end_date,
  pse_spe_filter = NULL,
  prestation_filter = NULL,
  analyse_couts = FALSE,
  dis_dtd_lag_months = 6,
  patients_ids_filter = NULL,
  output_table_name = NULL,
  conn = NULL
)
}
\arguments{
\item{start_date}{Date. La date de début de la période des consultations à
extraire.}

\item{end_date}{Date. La date de fin de la période des consultations à
extraire.}

\item{pse_spe_filter}{Character vector (Optionnel). Les codes spécialités des
médecins (référentiel \code{IR_SPE_V}) effectuant les consultations à extraire. Si
\code{pse_spe_filter} n'est pas fourni, les consultations de tous les spécialités
sont extraites. Défaut à \code{NULL}.}

\item{prestation_filter}{Character vector (Optionnel). Les codes des
prestations à extraire en norme B5 (colonne \code{PRS_NAT_REF}, référentiel
\code{IR_NAT_V}). Si \code{prestation_filter} n'est pas fourni, les consultations de
tous les prestations sont extraites. Les codes des prestations sont
disponibles sur la page \href{https://documentation-snds.health-data-hub.fr/snds/fiches/prestation.html}{"Cibler selon les natures de prestations"} de la documentation SNDS. Défaut à \code{NULL}.}

\item{analyse_couts}{Logical (Optionnel). Si \code{TRUE}, les filtres de qualité
liés aux coûts, écartant les actes en majorations sont ignorés. Utile pour
des analyses portant sur les coûts. Défaut à \code{FALSE}.}

\item{dis_dtd_lag_months}{Integer (Optionnel). Le nombre maximum de mois de
décalage de \code{FLX_DIS_DTD} par rapport à \verb{EXE_SOI DTD} pris en compte pour
récupérer les consultations. Défaut à 6 mois.}

\item{patients_ids_filter}{data frame (Optionnel). Un data frame contenant
les paires d'identifiants des patients pour lesquels les consultations
doivent être extraites. Les colonnes de ce data frame doivent être
\code{BEN_IDT_ANO}, \code{BEN_NIR_PSA} et \code{BEN_RNG_GEM}. Les BEN_NIR_PSA doivent être
tous les BEN_NIR_PSA associés
aux BEN_IDT_ANO fournis. Défaut à \code{NULL}.}

\item{output_table_name}{Character (Optionnel). Si fourni, les résultats
seront sauvegardés dans une table portant ce nom dans la base de données au
lieu d'être retournés sous forme de data frame. Si cette table existe déjà,
le programme s'arrête avec un message d'erreur. Défaut à \code{NULL}.}

\item{conn}{DBI connection (Optionnel). Une connexion à la base de données
Oracle. Par défaut, une connexion est établie avec oracle.}
}
\value{
Si \code{output_table_name} est \code{NULL}, retourne un data frame contenant
les consultations. Si \code{output_table_name} est fourni, sauvegarde les
résultats dans la table spécifiée dans Oracle et retourne \code{NULL} de manière
invisible.
Dans les deux cas les colonnes de la table de sortie sont :
\itemize{
\item \code{BEN_NIR_PSA} : Colonne présente uniquement si les identifiants
patients (\code{patients_ids_filter}) ne sont pas fournis. Identifiant SNDS,
aussi appelé pseudo-NIR.
\item \code{BEN_IDT_ANO} : Colonne présente uniquement si les identifiants
patients (\code{patients_ids_filter}) sont fournis. Numéro d’inscription
au répertoire (NIR) anonymisé.
\item \code{EXE_SOI_DTD} : Date de la consultation.
}
}
\description{
Cette fonction permet d'extraire les consultations dans le DCIR. Les
consultations dont les dates \code{EXE_SOI_DTD} sont comprises entre \code{start_date}
et \code{end_date} (incluses) sont extraites.
}
\details{
Le décalage de remontée des données est pris en compte en récupérant
également les consultations dont les dates \code{FLX_DIS_DTD} sont comprises dans
les \code{dis_dtd_lag_months} mois suivant \code{end_date}.

Si \code{patients_ids_filter} est fourni, seules les consultations pour les
patients dont les identifiants sont dans \code{patients_ids_filter} sont
extraites. Dans le cas contraire, les consultations de tous les patients sont
extraites.

Pour être à flux constant sur l'ensemble des années, il faut utiliser
\code{dis_dtd_lag_months = 27} Cela rallonge le temps d'extraction alors que
l'impact sur l'extraction est minime car \href{https://documentation-snds.health-data-hub.fr/snds/formation_snds/initiation/schema_relationnel_snds.html#_3-3-2-2-dates}{la Cnam estime que 99 \% des soins sont remontés à 6 mois}
c'est-à-dire pour \code{dis_dtd_lag_months = 6}.

Un guide sur l'activité des médecins libéraux est disponibles sur la page
\href{https://documentation-snds.health-data-hub.fr/snds/fiches/activite_medecins.html#contexte}{Activité des médecins libéraux}.
}
\examples{
\dontrun{
dispenses <- extract_consultations_erprsf(
  start_date = as.Date("2010-01-01"),
  end_date = as.Date("2010-01-03"),
  pse_spe_filter = c("0", "00", "36")
)
}
}
