% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_drug_erucdf.R
\name{extract_drug_erucdf}
\alias{extract_drug_erucdf}
\title{Extrait les dispensations de médicaments en accès précoces depuis le DCIR.}
\usage{
extract_drug_erucdf(
  start_date,
  end_date,
  ucd_codes_filter = NULL,
  patients_ids_filter = NULL,
  dis_dtd_lag_months = 6,
  sup_columns = NULL,
  output_table_name = NULL,
  conn = NULL,
  show_sql_query = FALSE
)
}
\arguments{
\item{start_date}{Début d'extraction (date de soin)}

\item{end_date}{Fin d'extraction (date de soin)}

\item{ucd_codes_filter}{Liste de codes ucd à extraire. Attention, les codes
UCD doivent être fournis au format UCD 7 caractères préfixé de 6 zéros :
"0000009419723". Ce format est celui utilisé dans la base de données.
Si NULL, extrait tous les codes.}

\item{patients_ids_filter}{data.frame (Optionnel). Un data.frame contenant
les paires d'identifiants des patients pour lesquels les délivrances de
médicaments doivent être extraites. Les colonnes de ce data.frame doivent
être "BEN_IDT_ANO" et "BEN_NIR_PSA". Les "BEN_NIR_PSA" doivent être tous
les "BEN_NIR_PSA" associés aux "BEN_IDT_ANO" fournis. Défaut à NULL.}

\item{dis_dtd_lag_months}{Integer (Optionnel). Le nombre maximum de mois de
décalage de FLX_DIS_DTD par rapport à EXE_SOI DTD pris en compte pour
récupérer les délivrances de médicaments. Défaut à 6 mois.}

\item{sup_columns}{Character vector (Optionnel). Les colonnes supplémentaires
à ajouter à la table de sortie. Défaut à NULL, donc aucune colonne ajoutée.}

\item{output_table_name}{Character (Optionnel). Si fourni, les résultats
seront sauvegardés dans une table portant ce nom dans la base de données au
lieu d'être retournés sous forme de data frame. Si la table existe déjà
dans la base oracle, alors le programme s'arrête en retournant une erreur.
Défaut à NULL.}

\item{conn}{DBI connection (Optionnel). Une connexion à la base de données
Oracle. Si non fournie, une connexion est établie par défaut. Défaut à
NULL.}

\item{show_sql_query}{Boolean (Optionnel). Affiche la requête SQL du premier
mois. Défaut à FALSE.}
}
\value{
Consommations individuelles d'accès précoces pour l'hospitalisation
privée et les rétrocessions hospitalières.

@examples
\dontrun{
 result <- extract_drug_erucdf(
   start_date = as.Date("2019-01-01"),
   end_date = as.Date("2019-12-31"),
   ucd_codes_filter = c("0000009419723"),
   output_table_name = "result_drug_erucdf",
   dis_dtd_lag_months = 6
 )
}*
}
\description{
Extrait les dispensations de médicaments en accès précoces réalisés en
hospitalisation privée (ex-OQN) ou en rétrocession hospitalière privée et
publique (ex-OQN et ex-DGF) pour une année donnée et une liste de codes UCD
donnée.
}
\details{
Les données sont filtrées par date de prestation (\code{EXE_SOI_DTD}), et par date
de flux (\code{FLX_DIS_DTD}) en regardant 7 mois au delà de la date de fin d'étude
\code{end_date} pour prendre en compte la durée de remontée de l'information. Les
données sont filtrées en ne conservant que les PRS_NAT_REF d'accès précoces
("3336", "3317", "3351", "3421").

NB: Les données sont extraites par date de flux, puis filtrées par date de
soin.  Elles sont extraites d'un bloc sur la période d'intérêt par contraste
avec les recommandations de la CNAM qui préconise d'extraire mois par mois
pour des raisons d'optimisation technique (éviter la saturation d'un
temporary space partagé entre utilisateurs). L'alternative serait de code une
fonction extrayant mois par mois (ie. "magic loop").

NB: La jointure avec la table établissement est faite avec un inner_join. On
ne garde que les AP préscrites en établissement.
}
\seealso{
Other extract: 
\code{\link{extract_consultations_erprsf}()},
\code{\link{extract_drug_dispenses}()},
\code{\link{extract_hospital_consultations}()},
\code{\link{extract_hospital_stays}()},
\code{\link{extract_long_term_disease}()}
}
\concept{extract}
