% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hospital_stays.R
\name{extract_hospital_stays}
\alias{extract_hospital_stays}
\title{Extraction des diagnostics des séjours hospitaliers (MCO).}
\usage{
extract_hospital_stays(
  start_date,
  end_date,
  dp_cim10_codes_filter = NULL,
  or_dr_with_same_codes_filter = FALSE,
  or_da_with_same_codes_filter = FALSE,
  and_da_with_other_codes_filter = FALSE,
  da_cim10_codes_filter = NULL,
  patients_ids_filter = NULL,
  output_table_name = NULL,
  conn = NULL
)
}
\arguments{
\item{start_date}{Date La date de début de la période sur laquelle extraire
les séjours.}

\item{end_date}{Date La date de fin de la période sur laquelle
extraire les séjours.}

\item{dp_cim10_codes_filter}{character vector
(Optionnel). Les codes CIM10 des diagnostics principaux à extraire. La
requête est effectuée par préfixe : Par exemple, si E12 est renseigné, tous
les codes commençant par E12 sont extraits. Défaut à \code{NULL}.}

\item{and_da_with_other_codes_filter}{logical (Optionnel). Indique si les
séjours avec des codes DA différents doivent être extraits. La requête est
effectuée par préfixe : Par exemple, si E12 est renseigné, tous les codes
commençant par E12 sont extraits. Défaut à \code{NULL}.}

\item{da_cim10_codes_filter}{character vector (Optionnel). Les codes CIM10
des diagnostics associés à extraire. La requête est effectuée par préfixe :
Par exemple, si E12 est renseigné, tous les codes commençant par E12 sont
extraits. Défaut à \code{NULL}.}

\item{patients_ids_filter}{data.frame (Optionnel). Un data.frame contenant
les paires d'identifiants des patients pour lesquels les consultations
doivent être extraites. Les colonnes de ce data.frame doivent être
\code{BEN_IDT_ANO} et \code{BEN_NIR_PSA} (en majuscules). Les "BEN_NIR_PSA" doivent
être tous les "BEN_NIR_PSA" associés aux "BEN_IDT_ANO" fournis. Si
\code{patients_ids} n'est pas fourni, les consultations de tous les patients sont
extraites. Défaut à \code{NULL}.}

\item{output_table_name}{character Le nom de la table de sortie dans la base
de données. Si \code{output_table_name} n'est pas fourni, une table de sortie
intermédiaire est créée. Défaut à \code{NULL}.}

\item{conn}{dbConnection La connexion à la base de données. Si \code{conn} n'est
pas fourni, une connexion à la base de données est initialisée. Défaut à
\code{NULL}.}

\item{or_dr_with_same_codes}{logical (Optionnel).Indique si les séjours avec
les mêmes codes DR doivent être extraits. La requête est effectuée par
préfixe : Par exemple, si E12 est renseigné, tous les codes commençant par
E12 sont extraits. Défaut à \code{NULL}.}

\item{or_da_with_same_codes}{logical (Optionnel). Indique si les séjours avec
les mêmes codes DA doivent être extraits. La requête est effectuée par
préfixe : Par exemple, si E12 est renseigné, tous les codes commençant par
E12 sont extraits. Défaut à \code{NULL}.}
}
\value{
Un data.frame contenant les séjours hospitaliers. Attention: Les
lignes des tables MCO B et C peuvent être dupliquées. Les colonnes sont les
suivantes :
\itemize{
\item \code{BEN_IDT_ANO} : Identifiant bénéficiaire anonymisé (seulement si
patient_ids non nul)
\item \code{NIR_ANO_17} : NIR anonymisé
\item \code{EXE_SOI_DTD} : Date de début du séjour hospitalier
\item \code{EXE_SOI_DTF} : Date de fin du séjour hospitalier
\item \code{ETA_NUM} : Numéro FINESS e-PMSI
\item \code{RSA_NUM} : N° d'index du RSA
\item \code{SEJ_NUM} : N° de séjour
\item \code{SEJ_NBJ} : Nombre de jours de séjour
\item \code{NBR_DGN} : Nombre de diagnostics associés significatifs
\item \code{NBR_RUM} : Nombre de RUM (unité médicales)
\item \code{NBR_ACT} : Nombre d'actes
\item \code{ENT_MOD} : Mode d'entrée
\item \code{ENT_PRV} : Provenance
\item \code{SOR_MOD} : Mode de sortie
\item \code{SOR_DES} : Destination
\item \code{DGN_PAL} : Diagnostic principal
\item \code{DGN_REL} : Diagnostic relié
\item \code{GRG_GHM} : Groupe homogène de malades
\item \code{BDI_DEP} : Département de résidence
\item \code{BDI_COD} : Code postal de résidence
\item \code{COD_SEX} : Sexe
\item \code{AGE_ANN} : Age en années
\item \code{AGE_JOU} : Age en jours
\item \code{DGN_PAL_UM} : Diagnostic principal des unité médicale
\item \code{DGN_REL_UM} : Diagnostic relié des unité médicale
\item \code{ASS_DGN} : Diagnostic associé
}
}
\description{
Cette fonction permet d'extraire les diagnostics des séjours
hospitaliers en MCO. Les diagnostics dont les dates \code{EXE_SOI_DTD} sont
comprises entre \code{start_date} et \code{end_date} sont extraits.
}
\details{
La sélection des séjours se fait à l'aide de filtres sur les
diagnostics:
\itemize{
\item Si \code{dp_cim10_codes_filter} est renseigné, seuls les séjours dont les
diagnostics principaux contiennent les codes CIM10 correspondants sont
extraits.
\item Si \code{or_dr_with_same_codes_filter} est renseigné, les séjours avec les codes
DR correspondants sont également extraits.
\item Si \code{or_da_with_same_codes_filter} est renseigné, les séjours avec les codes
DA correspondants sont également extraits.
\item Si \code{and_da_with_other_codes_filter} est renseigné, les séjours avec les
codes DA différents sont également extraits.
}

Tous les diagnostics principaux, reliés et associés sont extraits pour les
séjours sélectionnés.

La fonction joint les tables T_MCO\emph{B, T_MCO}C ensemble, puis joint
successivement à cette table "séjour" les tables T_MCO\emph{D et T_MCO}UM.
Finalement, les deux tables obtenues sont concaténées horizontalement. Il est
donc fréquent d'avoir des doublons concernant les colonnes des tables B et C
dans les lignes de la table résultante. Une explication détaillée et un
diagramme illustrant le fonctionnement retenu sont disponibles sur  \link[le
github du projet Scalpel]{https://github.com/X-DataInitiative/SCALPEL-Flattening/blob/DREES-104-DocFlattening/README_joins.md#the-pmsi-flattening}.
}
\examples{
\dontrun{
extract_hospital_stays(
  start_date =
    as.Date("2019-01-01"), end_date = as.Date("2019-12-31"), dp_cim10_codes =
    c("A00", "B00")
)  @export
}
}
