% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_long_term_disease.R
\name{extract_long_term_disease}
\alias{extract_long_term_disease}
\title{Extraction des Affections Longue Durée (ALD)}
\usage{
extract_long_term_disease(
  start_date = NULL,
  end_date = NULL,
  icd_cod_starts_with = NULL,
  ald_numbers = NULL,
  excl_etm_nat = c("11", "12", "13"),
  patients_ids = NULL,
  output_table_name = NULL,
  conn = NULL
)
}
\arguments{
\item{start_date}{Date La date de début de la période
sur laquelle extraire les ALD actives.}

\item{end_date}{Date La date de fin de la période
sur laquelle extraire les ALD actives.}

\item{icd_cod_starts_with}{character vector Un vecteur de codes
ICD 10. Si \code{icd_cod_starts_with} ou \code{ald_numbers} sont fournis,
seules les ALD associées à ces codes ICD 10 ou numéros d'ALD
sont extraites. Sinon, toutes les ALD actives sur la période
\link{start_date, end_date} sont extraites.}

\item{ald_numbers}{numeric vector Un vecteur de numéros d'ALD.
Si \code{icd_cod_starts_with} ou \code{ald_numbers} sont fournis,
seules les ALD associées à ces codes ICD 10 ou numéros d'ALD
sont extraites. Sinon, toutes les ALD actives sur la période
\link{start_date, end_date} sont extraites.}

\item{excl_etm_nat}{character vector Un vecteur de codes
IMB_ETM_NAT à exclure. Par défaut, les ALD de nature
11, 12 et 13 sont exclues car elles correspondent à des
exonérations pour accidents du travail ou maladies professionnelles.
Voir la fiche suivante de la documentation :
https://documentation-snds.health-data-hub.fr/snds/fiches/beneficiaires_ald.html
et notamment le Programme #1 pour la référence de ce filtre.}

\item{patients_ids}{data.frame Optionnel. Un data.frame contenant les
paires d'identifiants des patients pour lesquels les délivrances de
médicaments doivent être extraites. Les colonnes de ce data.frame
doivent être "ben_idt_ano" et "ben_nir_psa" (en minuscules). Les
"ben_nir_psa" doivent être tous les "ben_nir_psa" associés aux
"ben_idt_ano" fournis.}

\item{output_table_name}{Character Optionnel. Si fourni, les résultats seront
sauvegardés dans une table portant ce nom dans la base de données au lieu
d'être retournés sous forme de data frame.}

\item{conn}{DBI connection Une connexion à la base de données Oracle.
Si non fournie, une connexion est établie par défaut.}
}
\value{
Si output_table_name est NULL, retourne un data.frame contenant les
les ALDs actives sur la période. Si output_table_name est fourni,
sauvegarde les résultats dans la table spécifiée dans Oracle et
retourne NULL de manière invisible. Dans les deux cas les colonnes
de la table de sortie sont :
\itemize{
\item BEN_NIR_PSA : Colonne présente uniquement si les identifiants
patients (\code{patients_ids}) ne sont pas fournis. Identifiant SNDS,
ausi appelé pseudo-NIR.
\item BEN_IDT_ANO : Colonne présente uniquement si les identifiants
patients (\code{patients_ids}) sont fournis. Numéro d’inscription
au répertoire (NIR) anonymisé.
\item IMB_ALD_NUM : Le numéro de l'ALD
\item IMB_ALD_DTD : La date de début de l'ALD
\item IMB_ALD_DTF : La date de fin de l'ALD
\item IMB_ETM_NAT : La nature de l'ALD
\item MED_MTF_COD : Le code ICD 10 de la pathologie associée à l'ALD
}
}
\description{
Cette fonction permet d'extraire des ALD actives au
moins un jour sur une période donnée.
Les ALD dont l'intersection \link{IMB_ALD_DTD, IMB_ALD_DTF}
avec la période \link{start_date, end_date} n'est pas vide
sont extraites.
Si des codes ICD 10 ou des numéros d'ALD sont fournis,
seules les ALD associées à ces codes ICD 10 ou numéros
d'ALD sont extraites. Dans le cas contraire, toutes les
ALD sont extraites.
Si des identifiants de patients sont fournis, seules
les ALD associées à ces patients sont extraites. Dans
le cas contraire, les ALD de tous les patients sont extraites.
}
\examples{
\dontrun{
start_date <- as.Date("2010-01-01")
end_date <- as.Date("2010-01-03")
starts_with_codes <- c("N04A")

dispenses <- extract_drug_dispenses(
  start_date = start_date,
  end_date = end_date,
  starts_with_codes = starts_with_codes
)
}
}
