% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gather_table_stats}
\alias{gather_table_stats}
\title{Récupération des statistiques des tables}
\usage{
gather_table_stats(conn, table)
}
\arguments{
\item{conn}{Connexion à la base de données}

\item{table}{Chaine de caractère indiquant le nom d'une table}
}
\description{
Récupération des statistiques des tables
}
\references{
https://docs.oracle.com/en/database/oracle/oracle-database/19/arpls/DBMS_STATS.html#GUID-CA6A56B9-0540-45E9-B1D7-D78769B7714C
}
