% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{IS_PORTAIL}
\alias{IS_PORTAIL}
\title{Environnement d'exécution de sndsTools}
\format{
An object of class \code{logical} of length 1.
}
\usage{
IS_PORTAIL
}
\value{
Boolean. True si le package sndsTools est exécuté dans l'environnement de travail du Portail, False sinon.
}
\description{
Environnement d'exécution de sndsTools
}
\seealso{
Other constants: 
\code{\link{COLNAMES_DCIR_KEYS}}
}
\concept{constants}
\keyword{datasets}
