% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_synthetic_snds}
\alias{load_synthetic_snds}
\title{Charger les données synthétiques du SNDS}
\usage{
load_synthetic_snds()
}
\value{
Une liste de data frames, chacun correspondant à un fichier CSV dans
le zip.
}
\description{
Cette fonction charge les données synthétiques du SNDS à partir d'un fichier
zip contenu dans le répertoire \code{synthetic_data.zip}. Ce fichier zip contient
un fichier CSV pour chaque table du SDNS.
}
\details{
Le fichier zip \code{synthetic_data.zip} a été obtenu en téléchargeant \href{https://www.data.gouv.fr/fr/datasets/donnees-synthetiques-de-la-base-principales-du-systeme-national-des-donnees-de-sante/}{les données fictives du SNDS}
disponible sur datagouv.fr. Ces données ont été générées en 2019 grâce au
\href{https://gitlab.com/healthdatahub/applications-du-hdh/schema-snds}{schéma formel du SNDS} et au
projet
\href{https://gitlab.com/healthdatahub/se-former-au-snds/synthetic-generator/}{synthetic-generator}.

Ces données ne respectent pas les contraintes métier de la base de données
réelle. Par exemple, dans les données fictives, les dates de flux sont
générées aléatoirement par rapport aux dates de soin alors que dans la vraie
base de données, les dates de flux sont toujours postérieures aux dates de
soin.
}
\examples{
# Charger les données synthétiques
donnees_synthetiques <- load_synthetic_snds()
# Accéder à un data frame spécifique
#' df <- donnees_synthetiques[["ER_PRS_F"]]
#' # Afficher les premières lignes du data frame
#' print(head(df))
}
