% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hospital_stays.R
\name{normalize_column_number}
\alias{normalize_column_number}
\title{Normalise le nombre de colonnes pour les tables diagnostics d'unité médicale}
\usage{
normalize_column_number(df, column_prefix, max_columns_number)
}
\arguments{
\item{df}{data.frame Le data.frame à normaliser.}

\item{column_prefix}{character Le préfixe pour les colonnes à normaliser.}

\item{max_columns_number}{numeric Le nombre maximum de colonnes à conserver.}
}
\value{
data.frame Le data.frame normalisé.
}
\description{
Cette fonction limite le nombre de colonne à un nombre fixé par
l'utilisateur. Cela permet de se limiter aux \code{max_columns_number} diagnostics
seulement pour les diagnostics des unités médicales étant en nombre variable.
}
\examples{
df <- data.frame(
  DGN_PAL_UM_1 = "A", DGN_PAL_UM_2 = "B", DGN_PAL_UM_3 = "C"
)
normalize_column_number(df, "DGN_PAL_UM_", 2)
}
