% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_patient_id.R
\name{retrieve_psa}
\alias{retrieve_psa}
\title{Generic function retrieving patient identifiers}
\usage{
retrieve_psa(
  ben_table_name,
  start_key,
  check_arc_table = TRUE,
  output_table_name = NULL,
  conn = NULL
)
}
\arguments{
\item{ben_table_name}{Character Obligatoire. Nom de la table d'entrée
comprenant au moins la variable \code{BEN_IDT_ANO} ou \verb{BEN_NIR PSA}. Si la variable \code{BEN_RNG_GEM} est incluse, elle sera également
utilisée pour les jointures avec les référentiels.}

\item{start_key}{Character Obligatoire. Doit être égal à "BEN_IDT_ANO" ou
"BEN_NIR_PSA".}

\item{check_arc_table}{Logical Optionnel. Si TRUE (par défaut), les tables
\code{IR_BEN_R_ARC} sont également consultées pour la recherche des
\code{BEN_IDT_ANO} et des critères de sélection.}

\item{output_table_name}{Character Optionnel. Si fourni, les résultats seront
sauvegardés dans une table portant ce nom dans Oracle. Sinon la table en
sortie est retournée sous la forme d'un data.frame(/tibble).}

\item{conn}{DBI connection Optionnel Une connexion à la base de données
Oracle. Si non fournie, une connexion est établie par défaut.}
}
\description{
Refer to \code{retrieve_all_psa_from_idt} and \code{retrieve_all_psa_from_psa} for
details.
}
\seealso{
Other utils: 
\code{\link{check_output_table_name}()},
\code{\link{connect_duckdb}()},
\code{\link{connect_oracle}()},
\code{\link{constants_snds_tools}()},
\code{\link{create_table_from_query}()},
\code{\link{gather_table_stats}()},
\code{\link{get_first_non_archived_year}()},
\code{\link{insert_into_table_from_query}()},
\code{\link{retrieve_all_psa_from_idt}()},
\code{\link{retrieve_all_psa_from_psa}()}
}
\concept{utils}
