% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snds_table}
\alias{snds_table}
\title{Requêtage d'une table Oracle (SNDS ou ORAUSER).}
\usage{
snds_table(conn, table_name, profile = 108, schema = NULL)
}
\arguments{
\item{conn}{Connexion à la base de données.}

\item{table_name}{Nom de la table Oracle à explorer.}

\item{profile}{Nom de profil. Par défaut, profile = 108.}

\item{schema}{Si table hors du schéma standard, renseigner ici le schéma à
explorer. NULL par défaut.}
}
\value{
tbl_dbi. Exploration paresseuse (lazy query) d'une table Oracle (ORAUSER, ou SNDS).
}
\description{
Requêtage d'une table Oracle (SNDS ou ORAUSER).
}
\details{
La fonction détecte automatiquement les tables de l'ESND et de la
cartographie. Normalement, pas besoin d'utiliser l'argument schéma pour ces
tables.
}
\examples{
\dontrun{
# Connexion au SNDS
conn <- connect_oracle()

# Exploration paresseuse de la table ER_PRS_F
conn |>
 snds_table("ER_PRS_F") |>
 # Filtre sur la date de flux
 dplyr::filter(FLX_DIS_DTD == TO_DATE("2024-02-01", "yyyy-MM-dd"))

# Déconnexion du SNDS
DBI::dbDisconnect(conn)
}
}
\seealso{
Other utils: 
\code{\link{connect_duckdb}()},
\code{\link{connect_oracle}()},
\code{\link{create_table_from_query}()},
\code{\link{gather_table_stats}()},
\code{\link{get_first_non_archived_year}()},
\code{\link{insert_into_table_from_query}()},
\code{\link{retrieve_all_psa_from_idt}()},
\code{\link{retrieve_all_psa_from_psa}()}
}
\concept{utils}
