% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TNAME_FILTER_IR_PHA_R}
\alias{TNAME_FILTER_IR_PHA_R}
\title{Nom de la table temporaire pour les filtres ir_pha_r}
\format{
An object of class \code{character} of length 1.
}
\usage{
TNAME_FILTER_IR_PHA_R
}
\value{
A string with the name of the temporary table for ir_pha_r filters.
}
\description{
Nom de la table temporaire pour les filtres ir_pha_r
}
\seealso{
Other constants: 
\code{\link{TNAME_FILTER_PATIENTS}}
}
\concept{constants}
\keyword{datasets}
