% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TNAME_FILTER_PATIENTS}
\alias{TNAME_FILTER_PATIENTS}
\title{Nom de la table temporaire pour les IDs de patients}
\format{
An object of class \code{character} of length 1.
}
\usage{
TNAME_FILTER_PATIENTS
}
\value{
A string with the name of the temporary table for patient IDs.
}
\description{
Nom de la table temporaire pour les IDs de patients
}
\seealso{
Other constants: 
\code{\link{TNAME_FILTER_IR_PHA_R}}
}
\concept{constants}
\keyword{datasets}
