% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect_duckdb}
\alias{connect_duckdb}
\title{Initialisation de la connexion à la base de données duckdb.}
\usage{
connect_duckdb()
}
\value{
dbConnection Connexion à la base de données duckdb
}
\description{
Utilisation pour le testing uniquement. Si le code s'exécute en dehors du portail, il faut initier une connexion duckdb pour
effectuer les tests.
}
