% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_ace.R
\name{extract_hospital_consultations}
\alias{extract_hospital_consultations}
\title{Extrait les consultations externes à l'hôpital.}
\usage{
extract_hospital_consultations(
  start_date = NULL,
  end_date = NULL,
  spe_codes = NULL,
  ben_table_name = NULL,
  output_table_name = NULL,
  r_output_path = NULL
)
}
\arguments{
\item{start_date}{Date de début de la période}

\item{end_date}{Date de fin de la période}

\item{spe_codes}{Codes spécialités}

\item{ben_table_name}{Nom de la table BEN}

\item{output_table_name}{Nom de la table de sortie}

\item{r_output_path}{Chemin de sauvegarde des données R}
}
\value{
Tables MCO_FCSTC||MCO_CSC pour une année donnée
}
\description{
\itemize{
\item Sélectionne les codes prestations (ACT_COD) parmi : C et CS.
\item TODO: Implémente des filtres qualités sur les codes retours
}
}
